+++ src/LE/engine/input_system.cpp
@@ -38,7 +38,7 @@ void input_system::update_keystates()
 
   int x, y;
   p_curr_mouse_state = SDL_GetMouseState(&x, &y);
-  p_curr_mouse_pos.set((float)x, (float)y);
+  p_curr_mouse_pos.set((unsigned)x, (unsigned)y);
 
   // Quick hack
   if(p_mouse_wheel_set_this_update)
@@ -47,7 +47,7 @@ void input_system::update_keystates()
   }
   else
   {
-    p_mouse_wheel_delta = vec2::zero;
+    p_mouse_wheel_delta = uvec2::zero;
   }
 }
 
@@ -90,28 +90,28 @@ bool input_system::is_mouse_released(mouse_button button) const
       && (p_ignore_mouse == false);
 }
 
-vec2 const& input_system::get_mouse_pos() const
+uvec2 const& input_system::get_mouse_pos() const
 {
   return p_curr_mouse_pos;
 }
 
-vec2 const& input_system::get_prev_mouse_pos() const
+uvec2 const& input_system::get_prev_mouse_pos() const
 {
   return p_prev_mouse_pos;
 }
 
-vec2 input_system::get_delta_mouse_pos() const
+uvec2 input_system::get_delta_mouse_pos() const
 {
   return p_curr_mouse_pos - p_prev_mouse_pos;
 }
 
-void input_system::set_mouse_wheel_delta(vec2 const& new_delta)
+void input_system::set_mouse_wheel_delta(uvec2 const& new_delta)
 {
   p_mouse_wheel_set_this_update = true;
   p_mouse_wheel_delta = new_delta;
 }
 
-vec2 input_system::get_mouse_wheel_delta() const
+uvec2 input_system::get_mouse_wheel_delta() const
 {
   return p_mouse_wheel_delta;
 }
